+++ commonCharacters/commonCharacters.js
@@ -11,8 +11,36 @@
  */
 
 
-const commonCharacters = function commonCharacters(string1, string2) {
-  // TODO: Your code here!
+/**
+ * Get common characters between strings
+ * @param {...string} strings Strings to compare
+ */
+const commonCharacters = function commonCharacters() {
+  // Get all strings
+  const strings = Array.from(arguments);
+
+  if (strings.length === 0) {
+    return '';
+  }
+
+  const string1 = strings[0];
+
+  // Return unique common characters
+  return string1.split('').filter((char, index) => (
+    strings.every((string) => {
+      // Test string for common character
+      if (string.indexOf(char) >= 0) {
+        // Has character in common
+        if (string1.indexOf(char) === index) {
+          // Character is unique
+          return true;
+        }
+      }
+
+      // Character is not in common or not unique
+      return false;
+    })
+  )).join('');
 };
 
 if (window.DEBUG) {